+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Version 73
 ----------
 *in development*
 
+* 🌟 Discover: add new episodes screen with year and language filters, similar to popular shows screen.
 * 🔧 History: sort Trakt friends by latest activity first.
 * 🔨 Comments: do not display misleading "Could not modify data" message when commenting on a show.